SPT-1969: Renamed AuthorizationFunction
Renamed AuthorizationFunction to GetAuthorizeHandler, which now follows
our preferred naming strategy.
