+++ src/handlers/get-authorize-handler/get-authorize-handler.ts
@@ -1,4 +1,4 @@
-import { APIGatewayProxyEvent, APIGatewayProxyResult, Context } from "aws-lambda";
+import type { APIGatewayProxyEvent, APIGatewayProxyResult, Context } from "aws-lambda";
 import logger from "../../commons/logger";
 import { getRequiredEnvironment } from "../../commons/get-required-environment";
 import { URL } from "node:url";
